refactor(app): change the code of context API in a more Redux like style to make the code cleaner
Since the app is using the combination of useReducer and useContext to perform the state management,
we change the value object of the context API to be more Redux like style. Now we can destructure
the state and dispatch from the context API in each componment, which is more consistent.