refactor(app): replace action objects with action creators
Instead of creating the action objects mannually every where across the components, we use the
action creators to make the code more maintainable, testable and cleaner. In the case when we need
to modify the action, we no longer need to find all the places and update them all, we can instead
just modify the action creator.